Fix GPS configuration
Remove the built-in antenna delay in the timepulse configuration.
Antenna delay compensation is handled in the BroadPTP firmware or
can be set with get_gps_cable_delay.

Configure ZDA messages by default on all the ports. These should
be enabled by default but that doesn't appear to be the case in
all situations.

Bugfix: VRVDR-50402