Start ufispace-s5900-bsp-init only on S9500 platforms
vyatta-platform-name.service sets up a file with the name of
the platform in /var/run/vyatta/platform. Using the approprite
name as a condition in the service file enables systemd to skip
the service on platforms where it is not needed. This speeds up
system startup on all platforms. SIAD startup time is reduced
by about 1 second.

VRVDR-55611